+++ src/bittytax/conv/parsers/blockscout.py
@@ -1,15 +1,19 @@
 # -*- coding: utf-8 -*-
 # (c) Nano Nano Ltd 2020
 
+import sys
 from decimal import Decimal
-from typing import TYPE_CHECKING, Dict
+from typing import TYPE_CHECKING, Dict, List
 
+from colorama import Fore
 from typing_extensions import Unpack
 
 from ...bt_types import TrType
+from ...config import config
+from ...constants import WARNING
 from ..dataparser import DataParser, ParserArgs, ParserType
 from ..datarow import TxRawPos
-from ..exceptions import UnexpectedTypeError, UnknownCryptoassetError
+from ..exceptions import DataRowError, UnexpectedTypeError, UnknownCryptoassetError
 from ..out_record import TransactionOutRecord
 
 if TYPE_CHECKING:
@@ -18,7 +22,43 @@ if TYPE_CHECKING:
 WALLET = "Blockscout"
 
 
-def parse_blockscout(data_row: "DataRow", parser: DataParser, **kwargs: Unpack[ParserArgs]) -> None:
+def parse_blockscout(
+    data_rows: List["DataRow"], parser: DataParser, **kwargs: Unpack[ParserArgs]
+) -> None:
+
+    symbol = kwargs["cryptoasset"]
+    if not symbol:
+        sys.stderr.write(f"{WARNING} Cryptoasset cannot be identified\n")
+        sys.stderr.write(f"{Fore.RESET}Enter symbol: ")
+        symbol = input()
+        if not symbol:
+            raise UnknownCryptoassetError(kwargs["filename"], kwargs.get("worksheet", ""))
+
+    for data_row in data_rows:
+        if config.debug:
+            if parser.in_header_row_num is None:
+                raise RuntimeError("Missing in_header_row_num")
+
+            sys.stderr.write(
+                f"{Fore.YELLOW}conv: "
+                f" row[{parser.in_header_row_num + data_row.line_num}] {data_row}\n"
+            )
+
+        if data_row.parsed:
+            continue
+
+        try:
+            _parse_blockscout_row(data_row, parser, symbol)
+        except DataRowError as e:
+            data_row.failure = e
+        except (ValueError, ArithmeticError) as e:
+            if config.debug:
+                raise
+
+            data_row.failure = e
+
+
+def _parse_blockscout_row(data_row: "DataRow", parser: DataParser, symbol: str) -> None:
     row_dict = data_row.row_dict
     data_row.timestamp = DataParser.parse_timestamp(row_dict["UnixTimestamp"])
     data_row.tx_raw = TxRawPos(
@@ -27,9 +67,6 @@ def parse_blockscout(data_row: "DataRow", parser: DataParser, **kwargs: Unpack[P
         parser.in_header.index("ToAddress"),
     )
 
-    if not kwargs["cryptoasset"]:
-        raise UnknownCryptoassetError(kwargs["filename"], kwargs.get("worksheet", ""))
-
     quantity = Decimal(row_dict["Value"]) / 10**18
 
     if row_dict["Type"] == "IN":
@@ -37,7 +74,7 @@ def parse_blockscout(data_row: "DataRow", parser: DataParser, **kwargs: Unpack[P
             TrType.DEPOSIT,
             data_row.timestamp,
             buy_quantity=quantity,
-            buy_asset=kwargs["cryptoasset"],
+            buy_asset=symbol,
             wallet=_get_wallet(row_dict["ToAddress"]),
             note=_get_note(row_dict),
         )
@@ -51,9 +88,9 @@ def parse_blockscout(data_row: "DataRow", parser: DataParser, **kwargs: Unpack[P
             t_type,
             data_row.timestamp,
             sell_quantity=quantity,
-            sell_asset=kwargs["cryptoasset"],
+            sell_asset=symbol,
             fee_quantity=Decimal(row_dict["Fee"]) / 10**18,
-            fee_asset=kwargs["cryptoasset"],
+            fee_asset=symbol,
             wallet=_get_wallet(row_dict["FromAddress"]),
             note=_get_note(row_dict),
         )
@@ -124,7 +161,7 @@ DataParser(
         "TxDateClosingPrice",
     ],
     worksheet_name="Blockscout",
-    row_handler=parse_blockscout,
+    all_handler=parse_blockscout,
 )
 
 DataParser(